+++ common-client/src/main/java/de/maxhenkel/voicechat/gui/volume/AdjustVolumeList.java
@@ -79,7 +79,7 @@ public class AdjustVolumeList extends ListScreenListBase<VolumeEntry> {
                 if (volumeEntry instanceof PlayerVolumeEntry playerVolumeEntry) {
                     return playerVolumeEntry.getState() == null || !playerVolumeEntry.getState().getName().toLowerCase(Locale.ROOT).contains(filter);
                 } else if (volumeEntry instanceof CategoryVolumeEntry categoryVolumeEntry) {
-                    return !categoryVolumeEntry.getCategory().getName().toLowerCase(Locale.ROOT).contains(filter);
+                    return !categoryVolumeEntry.getCategory().getSearchName().toLowerCase(Locale.ROOT).contains(filter);
                 }
                 return true;
             });
@@ -93,7 +93,7 @@ public class AdjustVolumeList extends ListScreenListBase<VolumeEntry> {
                 }
             }
 
-            return volumeEntryToString(e1).compareToIgnoreCase(volumeEntryToString(e2));
+            return volumeEntryToSearchString(e1).compareToIgnoreCase(volumeEntryToSearchString(e2));
         });
         if (filter.isEmpty()) {
             filteredEntries.add(0, new PlayerVolumeEntry(null, screen));
@@ -101,11 +101,11 @@ public class AdjustVolumeList extends ListScreenListBase<VolumeEntry> {
         replaceEntries(filteredEntries);
     }
 
-    private String volumeEntryToString(VolumeEntry entry) {
+    private String volumeEntryToSearchString(VolumeEntry entry) {
         if (entry instanceof PlayerVolumeEntry playerVolumeEntry) {
             return playerVolumeEntry.getState() == null ? "" : playerVolumeEntry.getState().getName();
         } else if (entry instanceof CategoryVolumeEntry categoryVolumeEntry) {
-            return categoryVolumeEntry.getCategory().getName();
+            return categoryVolumeEntry.getCategory().getSearchName();
         }
         return "";
     }